スキルの時間制限を砂時計で表示できるようにした。

+++ Animation.h
@@ -6,14 +6,14 @@ public:
 	const char* textureName;	// 画像のパス
 	int textureId;				// 描画する画像
 	int gridAmount;				// アニメーションのコマ数
-	int sampleRate;				// 1秒間にアニメーションのコマを切り替える回数
+	float sampleRate;				// 1秒間にアニメーションのコマを切り替える回数
 	bool loopFlag;				// アニメーションがループするかどうか
 
 	// コンストラクタ
 	Animation(
 		const char* textureName,
 		int gridAmount = 1,
-		int sampleRate = 10,
+		float sampleRate = 10,
 		bool loopFlag = true
 	) :
 		textureName(textureName),
